requirements.txt
Flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.1
openai==1.93.0